feat(analytics-remote-config): remove idb (#855)
This is removing the use of IDB from remote config in favor of relying
on browser caching to help support remote config fetching. There were
definitely options to use IDB in other ways, but currently it's a bit
broken when trying to support multiple namespaces. Rather than trying
fix it, we'll see if just relying on the cache-control header and the
browser cache to solve the problem.