fix(session-replay-browser): fix heatmap click x/y
Use current scroll depth from the scroll watcher to
correctly track click x/y. Before we just relying on
the window scrolling element which is incorrect in cases
where the window is not the scrolling element. The
scroll events are correctly tracking depth so we're just
reusing that to calculate it for the clicks.

SR-1975